common/solvers: consolidate all algebra solvers into one reusable folder
Gather the scattered linear-algebra / optimization solvers into
src/common/solvers/ (joining the sparse Cholesky moved earlier):

- iop/choleski.h            -> common/solvers/choleski.h  (dense SPD Cholesky
                               + least-squares pseudo_solve)
- iop/gaussian_elimination.h-> common/solvers/gaussian_elimination.h
- iop/ashift_nmsimplex.c    -> common/solvers/nelder_mead_simplex.h  (the
                               Nelder-Mead simplex; now a self-contained
                               header: #pragma once + #ifndef-guarded
                               NMS_ALPHA/BETA/GAMMA defaults so ashift.c's
                               own values still win and other callers get
                               the standard coefficients)
- chart/thinplate.{c,h}     -> common/solvers/  (thin-plate spline; kept as
                               is, still references the chart tonecurve/
                               deltaE types -- the chart tool is not built)
- common/svd/{svd.h,QR_decomp.h,polar_decomposition.h,prova.c}
                            -> common/solvers/  (SVD, QR and polar
                               decomposition; common/svd/ removed)

All #include paths updated (highlights, toneequal, channelmixerrgb[_shared],
colorchecker, filmicrgb, ashift, chart/main, chart/common). src/CMakeLists.txt
HEADER_FILES glob extended with common/solvers/*.h. Pure relocation, no
solver code changed.

Behaviour-neutral: full RelWithDebInfo + GCC/clang Debug (-Werror) builds
green; harmonic sparse-cl self-test 2.1e-12, GT pk1synth/occluded identical
CPU+GPU.

Note: choleski.h and gaussian_elimination.h both define transpose_dot_matrix/
transpose_dot_vector/pseudo_solve* (float vs double); they are never included
in the same TU, but co-inclusion would now collide -- worth namespacing when
they next change.
