+++ src/CMakeLists.txt
@@ -223,7 +223,7 @@ FILE(GLOB SOURCE_FILES
   "views/dev_toolbox.c"
   "views/dev_backbuf.c"
   )
-FILE(GLOB HEADER_FILES "*.h" "common/*.h" "external/OpenCL/CL/*.h" "control/*.h" "iop/*.h" "libs/*.h" "views/*.h")
+FILE(GLOB HEADER_FILES "*.h" "common/*.h" "common/solvers/*.h" "external/OpenCL/CL/*.h" "control/*.h" "iop/*.h" "libs/*.h" "views/*.h")
 
 
 if(APPLE)
