+++ src/common/solvers/QR_decomp.h
