+++ src/common/solvers/thinplate.c
@@ -33,9 +33,9 @@
 #endif
 
 #include "common/darktable.h"
-#include "chart/thinplate.h"
+#include "common/solvers/thinplate.h"
 #include "chart/deltaE.h"
-#include "common/svd/svd.h"
+#include "common/solvers/svd.h"
 
 #include <assert.h>
 #include <float.h>
