+++ src/iop/channelmixerrgb.c
@@ -60,7 +60,7 @@
 #include "develop/imageop_math.h"
 #include "develop/openmp_maths.h"
 #include "dtgtk/drawingarea.h"
-#include "gaussian_elimination.h"
+#include "common/solvers/gaussian_elimination.h"
 #include "gui/color_picker_proxy.h"
 #include "gui/gtk.h"
 #include "gui/presets.h"
