+++ src/iop/filmicrgb.c
@@ -71,7 +71,7 @@
 #include "gui/color_picker_proxy.h"
 #include "gui/gtk.h"
 #include "gui/presets.h"
-#include "iop/gaussian_elimination.h"
+#include "common/solvers/gaussian_elimination.h"
 #include "iop/iop_api.h"
